+++ src/tomolog_cli/tomolog_2bm.py
@@ -145,147 +145,6 @@ class TomoLog2BM(TomoLog):
                 pass
         return proj
 
-    def read_recon(self):
-        log.info('Read reconstruction')
-        width = int(self.meta[self.width_key][0])
-        height = int(self.meta[self.height_key][0])
-        binning = int(self.meta[self.binning_key][0])
-        recon = []
-
-        if self.args.save_format == 'h5':
-            fname  = os.path.dirname(self.args.file_name)+'_rec/'+os.path.basename(self.args.file_name)[:-3]+'_rec.h5'
-            with h5py.File(fname,'r') as fid:
-                data = fid['exchange/recon']
-                h,w = data.shape[:2]
-                if self.args.idz == -1:
-                    self.args.idz = int(h//2)
-                if self.args.idy == -1:
-                    self.args.idy = int(w//2)
-                if self.args.idx == -1:
-                    self.args.idx = int(w//2)
-                if self.double_fov == True:
-                    binning_rec = np.log2(width//(w//2))
-                else:
-                    binning_rec = np.log2(width//(w))
-                x = data[:,:,self.args.idx]
-                y = data[:,self.args.idy]
-                z = data[self.args.idz]
-        else:
-            try:
-                basename = os.path.basename(self.args.file_name)[:-3]
-                dirname = os.path.dirname(self.args.file_name)
-                # set the correct prefix to find the reconstructions
-                rec_prefix = 'recon'
-
-                top = os.path.join(dirname+'_rec', basename+'_rec')
-                tiff_file_list = sorted(
-                    list(filter(lambda x: x.endswith(('.tif', '.tiff')), os.listdir(top))))
-                z_start = int(tiff_file_list[0].split('.')[0].split('_')[1])
-                z_end = int(tiff_file_list[-1].split('.')[0].split('_')[1]) + 1
-                height = z_end-z_start
-                fname_tmp = os.path.join(top, tiff_file_list[0])
-                # take size
-                tmp = utils.read_tiff(fname_tmp).copy()
-
-                if self.double_fov == True:
-                    width = width * 2
-                    binning_rec = 1
-                else:
-                    binning_rec = width//tmp.shape[0]
-
-                w = width//binning_rec
-                h = height
-                if self.args.idz == -1:
-                    self.args.idz = int(h//2)
-                if self.args.idy == -1:
-                    self.args.idy = int(w//2)+int(w//32)
-                if self.args.idx == -1:
-                    self.args.idx = int(w//2)-int(w//32)
-
-                z = utils.read_tiff(
-                    f'{dirname}_rec/{basename}_rec/{rec_prefix}_{self.args.idz:05}.tiff').copy()
-
-                y = utils.read_tiff(
-                    f'{dirname}_rec/{basename}_rec/{rec_prefix}_{self.args.idy:05}.tiff').copy()
-
-                x = utils.read_tiff(
-                    f'{dirname}_rec/{basename}_rec/{rec_prefix}_{self.args.idx:05}.tiff').copy()
-
-                # read x,y slices by lines
-                y = np.zeros((h, w), dtype='float32')
-                x = np.zeros((h, w), dtype='float32')
-
-                nthreads = 8
-                threads = []
-                lchunk = int(np.ceil((z_end-z_start)/nthreads))
-                lchunk = np.minimum(lchunk, np.int32(z_end-z_start-np.arange(nthreads)*lchunk))  # chunk sizes
-                for k in range(nthreads):
-                    read_proc = Thread(target=utils.read_tiff_part, args=(self.args, f'{dirname}_rec/{basename}_rec/{rec_prefix}', x, y, z_start, k*lchunk[0], lchunk[k]))
-                    threads.append(read_proc)
-                    read_proc.start()
-                for th in threads:
-                    th.join()
-                for j in range(z_start, z_end):
-                    zz = utils.read_tiff(
-                        f'{dirname}_rec/{basename}_rec/{rec_prefix}_{j:05}.tiff')
-                    y[j-z_start, :] = zz[self.args.idy]
-                    x[j-z_start, :] = zz[:, self.args.idx]
-
-                recon = [x, y, z]
-
-                self.binning_rec = binning_rec
-            except ZeroDivisionError:
-                log.error('Reconstructions for %s are larger than raw data image width. This is the case in a 0-360. Please use: --double-fov' % top)
-                log.warning('Skipping reconstruction')
-            except:
-                log.warning('Skipping reconstruction')
-        
-        return recon
-
-    def plot_recon(self, recon, fname):
-        log.info('Plot reconstruction')
-        fig = plt.figure(constrained_layout=True, figsize=(14, 12))
-        grid = fig.add_gridspec(3, 3, height_ratios=[1, 1, 1])
-        slices = ['x', 'y', 'z']
-        # autoadjust colorbar values according to a histogram
-
-        if self.args.min == self.args.max:
-            self.args.min, self.args.max = utils.find_min_max(
-                np.concatenate(recon))
-
-        sl = [self.args.idx, self.args.idy, self.args.idz]
-        tmp = literal_eval(self.args.zoom)
-        if not isinstance(tmp,list):
-            tmp = [tmp]
-        
-        zooms = tmp
-        log.info('Zooms selected %s' % zooms)
-        for j in range(3):
-            for k in range(3):
-                [s0,s1] = recon[k].shape
-                recon0 = recon[k][s0//2-s0//2//zooms[j]:s0//2+s0//2//zooms[j],s1//2-s1//2//zooms[j]:s1//2+s1//2//zooms[j]]
-                
-                recon0[0, 0] = self.args.max
-                recon0[0, 1] = self.args.min
-                recon0[recon0 > self.args.max] = self.args.max
-                recon0[recon0 < self.args.min] = self.args.min
-                ax = fig.add_subplot(grid[3*k+j])
-                im = ax.imshow(recon0, cmap='gray')
-                # Create scale bar
-                scalebar = ScaleBar(self.mct_resolution *
-                                    self.binning_rec, "um", length_fraction=0.25)
-                ax.add_artist(scalebar)
-                divider = make_axes_locatable(ax)
-                cax = divider.append_axes("right", size="5%", pad=0.1)
-                cb = plt.colorbar(im, cax=cax)
-                if j<2:
-                    cb.remove()
-                if j==0:
-                    ax.set_ylabel(f'slice {slices[k]}={sl[k]}', fontsize=18)
-        plt.savefig(fname, bbox_inches='tight', pad_inches=0, dpi=150)
-        plt.cla()
-        plt.close(fig)
-
     def publish_proj(self, presentation_id, page_id, proj):
         # 2-BM datasets may include both microCT data and a web camera image
         self.plot_projection(proj[0], self.file_name_proj0)
@@ -310,19 +169,3 @@ class TomoLog2BM(TomoLog):
                 presentation_id, page_id, 'Frame from the IP camera in the hutch', 160, 20, 10, 290, 8, 0)
         except:
             log.warning('No frame from the IP camera')
-
-    def publish_recon(self, presentation_id, page_id, recon):
-        if len(recon) == 3:
-            # publish reconstructions
-            self.plot_recon(recon, self.file_name_recon)
-            recon_url = filebin.upload(self.args, self.file_name_recon)
-            log.info('Publish reconstruction')
-            self.google_slide.create_image(
-                presentation_id, page_id, recon_url, 470, 400, 230, 5)
-            self.google_slide.create_textbox_with_text(
-                presentation_id, page_id, f'Reconstruction                                   Zoom {self.args.zoom}                                         ', 590, 20, 270, -10, 10, 0)
-
-            rec_line = self.read_rec_line()
-            self.google_slide.create_textbox_with_text(
-                presentation_id, page_id, rec_line, 1000, 20, 185, 391, 6, 0)
-
